+++ controllers/userController.js
@@ -2,19 +2,27 @@ const User = require("../models/User.js");
 
  
  const getAllUsers = async (req, res) => {
+  try {
  const users = await User.find({}).sort({ createdAt: -1 });
  res.status(200).json(users);
- };
+ }catch (error) {
+  res.status(500).json({ message: "Failed to retrieve Users" });
+ }};
 
  const createUser = async (req, res) => {
+  try{
     const newUser = await User.create({ ...req.body });
     res.status(201).json(newUser);
-  };
+  }catch (error) {
+    res.status(500).json({ message: "Failed to create User" });
+  }};
   
   
   const getUserById = async (req, res) => {
     const { userId } = req.params;
-  
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
     const user = await User.findById(userId);
     if (user) {
       res.status(200).json(user);
@@ -26,7 +34,9 @@ const User = require("../models/User.js");
   
   const updateUser = async (req, res) => {
     const { userId } = req.params;
-  
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
     const updatedUser = await Blog.findOneAndUpdate(
       { _id: userId },
       { ...req.body },
@@ -42,7 +52,10 @@ const User = require("../models/User.js");
  
   const deleteUser = async (req, res) => {
     const { userId } = req.params;
-  
+    
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
     const deletedUser = await User.findOneAndDelete({ _id: userId });
     if (deletedUser) {
       res.status(200).json({ message: "User deleted successfully" });
